still a long way to go, but refactored freetype to handle diagonal text MUCH better

+++ FreeTypeHelper.h
@@ -29,7 +29,7 @@ struct FontParameter //Represents all the information  required to call the writ
     std::string text = "";
     glm::vec4 rect = {0,0,0,0}; //if the width is negative, the height is assumed to be the font size
     glm::vec4 color = {0,0,0,1};
-    double angle = 0;
+    float angle = 0;
     float z = 0;
     Align align = LEFT;
     VertAlign vertAlign = UP;
